Moved client file to base since it's used by the NBViewer application, not specifically the URL provider. Removed some unused imports, updated copyright statements, and general tidying before tornado.options -> traitlets migration
