+++ nbviewer/providers/github/handlers.py
@@ -1,5 +1,5 @@
 #-----------------------------------------------------------------------------
-#  Copyright (C) 2013 The IPython Development Team
+#  Copyright (C) Jupyter Development Team
 #
 #  Distributed under the terms of the BSD License.  The full license is in
 #  the file COPYING, distributed as part of this software.
@@ -9,7 +9,6 @@ import os
 import json
 import mimetypes
 import re
-import asyncio
 
 from tornado import (
     web,
