+++ nbviewer/providers/github/tests/test_github.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 #-----------------------------------------------------------------------------
-#  Copyright (C) 2013 The IPython Development Team
+#  Copyright (C) Jupyter Development Team
 #
 #  Distributed under the terms of the BSD License.  The full license is in
 #  the file COPYING, distributed as part of this software.
