+++ nbviewer/tests/base.py
@@ -4,7 +4,7 @@ Derived from IPython.html notebook test case in 2.0
 """
 
 #-----------------------------------------------------------------------------
-#  Copyright (C) 2013 The IPython Development Team
+#  Copyright (C) Jupyter Development Team
 #
 #  Distributed under the terms of the BSD License.  The full license is in
 #  the file COPYING, distributed as part of this software.
