+++ nbviewer/tests/test_security.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 #-----------------------------------------------------------------------------
-#  Copyright (C) 2015 The Jupyter Development Team
+#  Copyright (C) Jupyter Development Team
 #
 #  Distributed under the terms of the BSD License.  The full license is in
 #  the file COPYING, distributed as part of this software.
